GetConfig in pipeline package (#358)
